Add `_setup_env` bash function
Copy functionality from `git-tmux` to this function. With this when
opening a new bash session then if it is in a git repo it will activate
the corresponding virtualenv if there is one, source the chain of
`.project_env` files, rename the tmux window and set the `PROJECT_NAME`
variable so bash history is saved in the right place.