Map caps lock to escape
This is a bit tricky to automate. Modifier mappings are stored per keyboard in defaults. In order to find the correct key to store the mapping the keyboard needs to be identified first.

This can be done with ioreg. Special care needs to be taken as intel macs use USB keyboards internally, apple silicon macs use SPI keyboard, so these reside under different classes in ioreg. Ioreg can output in XML, so we can find the id using xpath querying. Since different keyboards reside under different classes, the query is for all devices, making it quite slow...

Once the product id is found using ioreg, we first check the current value and if it is incorrect then set it. This has to be done this way because we need to write the key in XML format, so that the mappings are stored in the correct integer type, not as strings which is the default. When storing in XML format the diff checking doesn't work correctly, thus the check beforehand.